+++ docs/release_notes.rst
@@ -25,8 +25,9 @@ SQLite and Parquet Outputs
   :issue:`1176,806`.
 * Data types, specified value constraints, and the uniqueness / non-null
   constraints on primary keys are validated during insertion into the SQLite DB.
-* The PUDL ETL CLI :mod:`pudl.cli` now has flags to activate various constraint
-  checks including ``--check-foreign-keys``, ``--check-types``, and ``--check-values``.
+* The PUDL ETL CLI :mod:`pudl.cli` now has flags to toggle various constraint
+  checks including ``--ignore-foreign-key-constraints``,
+  ``--ignore-type-constraints``, and ``--ignore-value-constraints``.
 
 New Metadata System
 ^^^^^^^^^^^^^^^^^^^